RF: move protocols to separate file
